feat(settings): Voice Providers panel — accessible from /x/settings
Extracts the CRUD UI from the standalone /x/settings/voice-providers
page into <VoiceProvidersPanel /> so it also renders inside the unified
Settings hub under the Communications category.

- New: components/settings/VoiceProvidersPanel.tsx (shared body)
- Standalone page now thin wraps the panel with a page heading
- SettingsClient registers it as 'voice_providers' panel
- SettingsSidebar gains PhoneCall icon
- globals.css adds the .hf-archive-toggle-row helpers

So clicking Settings in the sidebar → Voice Providers (Communications)
gets you the full CRUD without the standalone URL.
