+++ apps/admin/app/globals.css
@@ -7690,6 +7690,18 @@ html.light {
   opacity: 0.7;
 }
 
+/* Voice Providers — "Show archived" toggle row */
+.hf-archive-toggle-row {
+  justify-content: flex-end;
+  margin-bottom: 12px;
+}
+.hf-archive-toggle-label {
+  display: inline-flex;
+  align-items: center;
+  gap: 8px;
+  cursor: pointer;
+}
+
 /* Progress track thin (6px) */
 .hf-progress-track-thin {
   height: 6px;
